+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 - **angular-formly** has been upgraded from 8.2.0 to **8.4.1**
 - **bootstrap-sass** has been upgraded from 3.3.6 to **3.3.7**
 - **ng-tasty** has been upgraded from 0.5.8 to **0.6.1**
+- **angular-bootstrap** has been upgraded from 1.1.2 to **2.5.0**
 - **checklist-model** has been removed
 
 # 5.0.0 / 2018-05-04